+++ packages/react/src/useMount.ts
@@ -1,6 +1,17 @@
 import ComponentSetup from './ComponentSetup';
+import useContextType from './useContextType';
+import { useComponentLifeCycle, usePageLifeCycle } from '@goldfishjs/composition-api';
 
 export default function useMount(fn: () => void) {
-  const setup = ComponentSetup.getCurrent<ComponentSetup>();
-  setup.mountFns.push(fn);
+  const contextType = useContextType();
+  if (contextType === 'react') {
+    const setup = ComponentSetup.getCurrent<ComponentSetup>();
+    setup.mountFns.push(fn);
+  } else if (contextType === 'component') {
+    useComponentLifeCycle('didMount', fn);
+  } else if (contextType === 'page') {
+    usePageLifeCycle('onReady', fn);
+  } else {
+    throw new Error(`The useMount can not be used in ${contextType} setup flow.`);
+  }
 }